# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SignupRequestType =
  (typeof SignupRequestType)[keyof typeof SignupRequestType];

export const SignupRequestType = {
  organization: "organization",
  individual: "individual",
} as const;

export interface SignupRequest {
  type: SignupRequestType;
  name: string;
  contactName?: string;
  email: string;
  password: string;
  country?: string;
  state?: string;
  city?: string;
}

export interface SigninRequest {
  email: string;
  password: string;
}

export type ProfileType = (typeof ProfileType)[keyof typeof ProfileType];

export const ProfileType = {
  organization: "organization",
  individual: "individual",
} as const;

export interface Profile {
  type: ProfileType;
  name: string;
  contactName?: string;
  email: string;
  phone?: string;
  country?: string;
  state?: string | null;
  city?: string;
  address?: string;
  postalCode?: string;
  gstin?: string;
  pan?: string;
  taxId?: string;
  website?: string;
  industry?: string;
  password?: string;
  logo?: string;
  signature?: string;
  stamp?: string;
}

export interface Preferences {
  currency?: string;
  invoiceFormat?: string;
  invoicePrefix?: string;
  taxMode?: string;
  defaultGstRate?: number;
  defaultTerms?: string;
  defaultNotes?: string;
  paymentBankName?: string;
  paymentAccount?: string;
  paymentIfsc?: string;
  paymentUpi?: string;
}

export interface InvoiceData {
  [key: string]: unknown;
}