# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: profile
    description: Profile operations
  - name: preferences
    description: Preferences operations
  - name: invoices
    description: Invoices operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      summary: Sign up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: Signed up successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /auth/signin:
    post:
      operationId: signIn
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SigninRequest"
      responses:
        "200":
          description: Signed in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get profile
      responses:
        "200":
          description: Profile retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    post:
      operationId: saveProfile
      tags: [profile]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Profile"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /preferences:
    get:
      operationId: getPreferences
      tags: [preferences]
      summary: Get preferences
      responses:
        "200":
          description: Preferences retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
    post:
      operationId: savePreferences
      tags: [preferences]
      summary: Update preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Preferences"
      responses:
        "200":
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
  /invoices/draft:
    get:
      operationId: getDraftInvoice
      tags: [invoices]
      summary: Get draft invoice
      responses:
        "200":
          description: Draft invoice retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceData"
    post:
      operationId: saveDraftInvoice
      tags: [invoices]
      summary: Save draft invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceData"
      responses:
        "200":
          description: Draft invoice saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceData"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SignupRequest:
      type: object
      properties:
        type:
          type: string
          enum: [organization, individual]
        name:
          type: string
        contactName:
          type: string
        email:
          type: string
        password:
          type: string
        country:
          type: string
        state:
          type: string
        city:
          type: string
      required:
        - type
        - name
        - email
        - password
    SigninRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    Profile:
      type: object
      properties:
        type:
          type: string
          enum: [organization, individual]
        name:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        state:
          type: string
          nullable: true
        city:
          type: string
        address:
          type: string
        postalCode:
          type: string
        gstin:
          type: string
        pan:
          type: string
        taxId:
          type: string
        website:
          type: string
        industry:
          type: string
        password:
          type: string
        logo:
          type: string
        signature:
          type: string
        stamp:
          type: string
      required:
        - type
        - name
        - email
    Preferences:
      type: object
      properties:
        currency:
          type: string
        invoiceFormat:
          type: string
        invoicePrefix:
          type: string
        taxMode:
          type: string
        defaultGstRate:
          type: integer
        defaultTerms:
          type: string
        defaultNotes:
          type: string
        paymentBankName:
          type: string
        paymentAccount:
          type: string
        paymentIfsc:
          type: string
        paymentUpi:
          type: string
    InvoiceData:
      type: object
      additionalProperties: true