# lib/api-zod/src/generated/types/preferences.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Preferences {
  currency?: string;
  invoiceFormat?: string;
  invoicePrefix?: string;
  taxMode?: string;
  defaultGstRate?: number;
  defaultTerms?: string;
  defaultNotes?: string;
  paymentBankName?: string;
  paymentAccount?: string;
  paymentIfsc?: string;
  paymentUpi?: string;
}