# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileType } from "./profileType";

export interface Profile {
  type: ProfileType;
  name: string;
  contactName?: string;
  email: string;
  phone?: string;
  country?: string;
  state?: string | null;
  city?: string;
  address?: string;
  postalCode?: string;
  gstin?: string;
  pan?: string;
  taxId?: string;
  website?: string;
  industry?: string;
  password?: string;
  logo?: string;
  signature?: string;
  stamp?: string;
}