# lib/api-zod/src/generated/types/signupRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SignupRequestType } from "./signupRequestType";

export interface SignupRequest {
  type: SignupRequestType;
  name: string;
  contactName?: string;
  email: string;
  password: string;
  country?: string;
  state?: string;
  city?: string;
}